Modify object selection while loading
When the object is initialized, it has no connection with the actual service.
After it's been tried to be accesses, it loads some information that gets
copied to the current objects attributes.
While selecting objects to be copied, it would copy the first object found
and would exit. Therefore, max version selection woudn't be performed.
It's been fixed to look over all versions of relevant objects.