Add filtering serivce objects by version filters
The problem: Service->SDC_Resource->SDC inheritance.
The version filter is specified in Service, the method taking the version filter
from Service is in SDC. SDC inherited by children other than Service will
not retrieve the filter that's not there and it'll result in errors.

The solution: check if the filter exists first to avoid further evaluation
(short-curcuit evaluation) as only Service for now needs it.